pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langflow-mcp-tool"
version = "0.1.0"
description = "MCP Tool 用於 Langflow 文檔問答系統"
requires-python = ">=3.10"
authors = [
    {name = "User", email = "user@example.com"}
]
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "mcp[cli]>=1.3.0",
    "typer>=0.9.0",
    "uvicorn>=0.22.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
langflow-mcp = "langflow_mcp_tool.cli:cli"

[tool.setuptools]
packages = ["langflow_mcp_tool"] 
